[Chore] psd_loader.py 파일 data_load.py 와 병합
